changed int data type to ssize_t in 0-read_textfile.c

+++ 0x15-file_io/0-read_textfile.c
@@ -9,7 +9,7 @@
  */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	int fd, n;
+	ssize_t fd, n;
 	char *buffer;
 
 	if (filename == NULL)
